requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
requests==2.31.0
SpeechRecognition==3.10.0
pyttsx3==2.90
pydub==0.25.1
Werkzeug==2.3.7
gunicorn==21.2.0
PyAudio==0.2.11
google-generativeai==0.3.2